Add support for TypedArrays in User Scripts using @foxglove/schemas (#6890)
**User-Facing Changes**
Fixed a type incompatibility issue with User Scripts that import from
@foxglove/schemas.

**Description**
Fixes #6821 
Fixes FG-4976
Depends on https://github.com/foxglove/schemas/pull/128
Also updated JSON message datatype extraction to treat integer arrays as
`Uint32Array` rather than `Float64Array`; this fixed a similar issue as
#6821 with an example json-based mcap I had lying around.

+++ packages/studio-base/src/components/TopicList/getMessagePathSearchItems.test.ts
@@ -295,7 +295,7 @@ describe("getMessagePathSearchItems", () => {
           "suffix": {
             "isLeaf": true,
             "pathSuffix": ".circles[:].timestamp.sec",
-            "type": "float64",
+            "type": "uint32",
           },
           "topic": {
             "name": "annotations",
@@ -308,7 +308,7 @@ describe("getMessagePathSearchItems", () => {
           "suffix": {
             "isLeaf": true,
             "pathSuffix": ".circles[:].timestamp.nsec",
-            "type": "float64",
+            "type": "uint32",
           },
           "topic": {
             "name": "annotations",
@@ -542,7 +542,7 @@ describe("getMessagePathSearchItems", () => {
           "suffix": {
             "isLeaf": true,
             "pathSuffix": ".points[:].timestamp.sec",
-            "type": "float64",
+            "type": "uint32",
           },
           "topic": {
             "name": "annotations",
@@ -555,7 +555,7 @@ describe("getMessagePathSearchItems", () => {
           "suffix": {
             "isLeaf": true,
             "pathSuffix": ".points[:].timestamp.nsec",
-            "type": "float64",
+            "type": "uint32",
           },
           "topic": {
             "name": "annotations",
@@ -854,7 +854,7 @@ describe("getMessagePathSearchItems", () => {
           "suffix": {
             "isLeaf": true,
             "pathSuffix": ".texts[:].timestamp.sec",
-            "type": "float64",
+            "type": "uint32",
           },
           "topic": {
             "name": "annotations",
@@ -867,7 +867,7 @@ describe("getMessagePathSearchItems", () => {
           "suffix": {
             "isLeaf": true,
             "pathSuffix": ".texts[:].timestamp.nsec",
-            "type": "float64",
+            "type": "uint32",
           },
           "topic": {
             "name": "annotations",
